Added respons when handling products as admin

+++ html/actions/delete_product.php
@@ -1,22 +1,28 @@
 <?php
-	include "../functions.php";
-	$con = connect();
-	$id = secure_str($_GET["id"]);
+    session_start();
+    if(isset($_SESSION['admin'])){
+        include "../functions.php";
+        $con = connect();
+        $id = secure_str($_GET["id"]);
 
-	// deletes the product
-	$query = "DELETE FROM product WHERE product_id = '$id'";
+        // deletes the product
+        $query = "DELETE FROM product WHERE product_id = '$id'";
 
-	$select = mysqli_query($con, $query) or die (mysqli_error($con));
+        $select = mysqli_query($con, $query) or die (mysqli_error($con));
 
-	// delets all the key_features of this product
-	$query = "DELETE FROM key_feature WHERE product_id = '$id'";
+        // delets all the key_features of this product
+        $query = "DELETE FROM key_feature WHERE product_id = '$id'";
 
-	$select = mysqli_query($con, $query) or die (mysqli_error($con));
-	
-	// deleted all the tech_table of this product
-	$query = "DELETE FROM tech_table_row WHERE product_id = '$id'";
+        $select = mysqli_query($con, $query) or die (mysqli_error($con));
+        
+        // deleted all the tech_table of this product
+        $query = "DELETE FROM tech_table_row WHERE product_id = '$id'";
 
-	$select = mysqli_query($con, $query) or die (mysqli_error($con));
+        $select = mysqli_query($con, $query) or die (mysqli_error($con));
 
-	header("Location: ../admin.php");
+        header("Location: ../admin?message=Product has been deleted");
+    }
+    else{
+        header("Location: ../index");
+    }
 ?>
